refactor : 서비스에 맞는 토큰을 반환하는 클래스명 변경
